Stop decoding all %2e's in paths
This basically reverts bee5ad8041adfe6fc676c527bbc3f3cf4562ef67 as it
did not end up shipping in Firefox and the tentative conclusion is that
percent-decoding in URLs while parsing creates more trouble than it’s
worth.

We’ll still handle %2e when it’s part of a single-dot or double-dot
path segment.

Fixes #87.